requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0
ipdb==0.13.13
debugpy==1.8.0
loguru==0.7.2
rich==13.7.0
flask==3.0.0
flask-assets==2.1.0 